deprecate format_template
