+++ changes/345.removal.rst
@@ -1,0 +1,1 @@
+Deprecate ``stpipe.format_template``. Use fstrings instead.
